requirements.txt
torch==2.0.0
pandas==1.5.3
sentence_transformers==0.4.1.2
google.generativeai==0.1.0